+++ src/main/java/org/team1540/robot2019/datastructures/utils/UnitsUtils.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.utils;
+package org.team1540.robot2019.datastructures.utils;
 
 public class UnitsUtils {
