+++ src/main/java/org/team1540/robot2019/drivecontrol/commands/UDPVelocityTwistDrive.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.vision.commands;
+package org.team1540.robot2019.drivecontrol.commands;
 
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
@@ -10,8 +10,8 @@ import org.team1540.robot2019.Robot;
 import org.team1540.robot2019.Tuning;
 import org.team1540.robot2019.datastructures.twod.Transform2D;
 import org.team1540.robot2019.datastructures.twod.Twist2D;
+import org.team1540.robot2019.datastructures.utils.TrigUtils;
 import org.team1540.robot2019.utils.TankDriveTwist2DInput;
-import org.team1540.robot2019.utils.TrigUtils;
 import org.team1540.rooster.drive.pipeline.CTREOutput;
 import org.team1540.rooster.drive.pipeline.FeedForwardProcessor;
 import org.team1540.rooster.drive.pipeline.UnitScaler;
