+++ src/main/java/org/team1540/robot2019/vision/LimelightLocalization.java
@@ -1,7 +1,8 @@
-package org.team1540.robot2019.utils;
+package org.team1540.robot2019.vision;
 
 import org.apache.commons.math3.geometry.euclidean.threed.Rotation;
 import org.apache.commons.math3.geometry.euclidean.threed.RotationConvention;
+import org.apache.commons.math3.geometry.euclidean.threed.RotationOrder;
 import org.apache.commons.math3.geometry.euclidean.threed.Vector3D;
 import org.apache.commons.math3.geometry.euclidean.twod.Vector2D;
 import org.team1540.robot2019.Tuning;
@@ -45,7 +46,8 @@ public class LimelightLocalization {
         Rotation cameraTilt = new Rotation(Vector3D.PLUS_J, CAMERA_TILT, RotationConvention.FRAME_TRANSFORM);
         Rotation cameraRoll = new Rotation(Vector3D.PLUS_I, CAMERA_ROLL, RotationConvention.FRAME_TRANSFORM);
 
-        Rotation cameraRotation = cameraTilt.applyTo(cameraRoll);
+        Rotation cameraRotation = new Rotation(RotationOrder.XYZ, RotationConvention.FRAME_TRANSFORM, CAMERA_ROLL, CAMERA_TILT, 0);
+//        Rotation cameraRotation = cameraTilt.applyTo(cameraRoll);
         baseLinkToVisionTarget = DualVisionTargetLocalizationUtils
             .poseFromTwoCamPoints(point0, point1, PLANE_HEIGHT, CAMERA_POSITION, cameraRotation, Tuning.LIMELIGHT_HORIZONTAL_FOV, Tuning.LIMELIGHT_VERTICAL_FOV);
